ticket VS-1260 adding the ability to decrypt tls_1.2 with a keylogger with using wireshark decryption routines  - enable via new option ssl_tls_12_mode = ws (default is dssl) --- adding verbose parameter disable_unlink_qfile

+++ sql_db.cpp
@@ -4332,7 +4332,9 @@ bool MySqlStore::loadFromQFile(const char *filename, int id_main, bool onlyCheck
 	} while(true);
 	#endif
 	if(!onlyCheck) {
-		unlink(filename);
+		if(!sverb.disable_unlink_qfile) {
+			unlink(filename);
+		}
 		//rename(filename, find_and_replace(filename, "qoq", "_qoq").c_str());
 	}
 	if(sverb.qfiles) {
